requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
bleach==6.2.0; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.3; python_version >= '3.7'
cryptography==45.0.6; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==5.2.5; python_version >= '3.10'
django-filter==25.1; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
djoser==2.3.3; python_version >= '3.9' and python_version < '4.0'
idna==3.10; python_version >= '3.6'
mysqlclient==2.2.7; python_version >= '3.8'
oauthlib==3.3.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-decouple==3.8
python-stdnum==2.1; python_version >= '3.8'
python3-openid==3.2.0
requests==2.32.5; python_version >= '3.9'
requests-oauthlib==2.0.0; python_version >= '3.4'
social-auth-app-django==5.5.1; python_version >= '3.10'
social-auth-core==4.7.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
urllib3==2.5.0; python_version >= '3.9'
webencodings==0.5.1
